More Google IMA examples, WIP...
