+++ .idea/dictionaries/trk.xml
@@ -1,6 +1,7 @@
 <component name="ProjectDictionaryState">
   <dictionary name="trk">
     <words>
+      <w>aaid</w>
       <w>adid</w>
       <w>amazonaws</w>
       <w>appfw</w>
