+++ app/src/main/res/layout/activity_ad_on_top.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent"
+	android:orientation="vertical">
+
+	<fi.finwe.orion360.sdk.pro.view.OrionViewContainer
+		android:id="@+id/orion_view_container"
+		android:layout_width="match_parent"
+		android:layout_height="match_parent" />
+
+	<com.google.android.exoplayer2.ui.StyledPlayerView
+		android:id="@+id/exoplayer_styled_player_view"
+		android:layout_width="match_parent"
+		android:layout_height="400dp" />
+
+</FrameLayout>
